refactor(executor): decompose predicate.rs into focused modules (#4867)
Split the 1745-line predicate.rs into focused modules by predicate type:

- `mod.rs` (75 lines): Common types and re-exports
- `range.rs` (358 lines): Range predicate extraction and merging
- `in_list.rs` (400 lines): IN predicate extraction and main entry point
- `composite.rs` (605 lines): Multi-column composite index predicates
- `expression_index.rs` (386 lines): Expression-based index support

Each module now handles one predicate category with clear responsibilities.
All 2258 tests pass.

Closes #4867
